+++ packages/nitro/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @vercel/workflow-nitro
 
+## 4.0.1
+
+### Patch Changes
+
+- 4ca9a3e: Add workflow/nitro
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [b472bb1]
+- Updated dependencies [4ca9a3e]
+- Updated dependencies [57419e5]
+  - @workflow/cli@4.0.1
+  - @workflow/core@4.0.1
+  - @workflow/swc-plugin@4.0.1
+
+## 4.0.1-beta.3
+
+### Patch Changes
+
+- Updated dependencies [b472bb1]
+- Updated dependencies [57419e5]
+  - @workflow/cli@4.0.1-beta.3
+  - @workflow/core@4.0.1-beta.3
+
 ## 4.0.1-beta.2
 
 ### Patch Changes